requirements.txt
aiogram==2.25.2
python-dotenv==1.0.1
apscheduler>=3.10.0
